+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add new command ImportDatabase to generate RDMP metadata for an existing SMI-type DB
+
 ## [7.1.3] 2024-12-02
 
 - Bump RDMP from 8.3.1 to 8.4.0